+++ uikit-custom-sample/src/main/java/com/sendbird/uikit/customsample/HomeActivity.java
@@ -1,13 +1,24 @@
 package com.sendbird.uikit.customsample;
 
+import android.Manifest;
 import android.app.NotificationManager;
 import android.content.Context;
 import android.content.Intent;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.provider.Settings;
 import android.view.View;
 
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.PermissionChecker;
 import androidx.databinding.DataBindingUtil;
 
 import com.sendbird.android.GroupChannelTotalUnreadMessageCountParams;
@@ -22,15 +33,22 @@ import com.sendbird.uikit.customsample.groupchannel.GroupChannelMainActivity;
 import com.sendbird.uikit.customsample.openchannel.OpenChannelMainActivity;
 import com.sendbird.uikit.customsample.utils.PreferenceUtils;
 import com.sendbird.uikit.customsample.utils.PushUtils;
+import com.sendbird.uikit.utils.ContextUtils;
 import com.sendbird.uikit.widgets.WaitingDialog;
 
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 
 
 public class HomeActivity extends AppCompatActivity {
     private ActivityHomeBinding binding;
     private static final String USER_EVENT_HANDLER_KEY = "USER_EVENT_HANDLER_KEY" + System.currentTimeMillis();
+
+    @NonNull
+    private final ActivityResultLauncher<String> requestPermissionLauncher = registerForActivityResult(new ActivityResultContracts.RequestPermission(), isGranted -> {});
+    @NonNull
+    private final ActivityResultLauncher<Intent> appSettingLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), intent -> {});
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -45,6 +63,18 @@ public class HomeActivity extends AppCompatActivity {
 
         binding.tvUnreadCount.setTextAppearance(this, R.style.SendbirdCaption3OnDark01);
         binding.tvUnreadCount.setBackgroundResource(R.drawable.shape_badge_background);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            final String permission = Manifest.permission.POST_NOTIFICATIONS;
+            if (ContextCompat.checkSelfPermission(this, permission) == PermissionChecker.PERMISSION_GRANTED) {
+                return;
+            }
+            if (ActivityCompat.shouldShowRequestPermissionRationale(this, permission)) {
+                showPermissionRationalePopup();
+                return;
+            }
+            requestPermissionLauncher.launch(permission);
+        }
     }
 
     @Override
@@ -119,5 +149,23 @@ public class HomeActivity extends AppCompatActivity {
                 WaitingDialog.dismiss();
             }
         });
+    }
+
+    private void showPermissionRationalePopup() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setTitle(getString(com.sendbird.uikit.R.string.sb_text_dialog_permission_title));
+        builder.setMessage(String.format(Locale.US, getString(R.string.sb_text_need_to_allow_permission_notification), ContextUtils.getApplicationName(this)));
+        builder.setPositiveButton(com.sendbird.uikit.R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
+            Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.addCategory(Intent.CATEGORY_DEFAULT);
+            intent.setData(Uri.parse("package:" + getPackageName()));
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
+            intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
+            appSettingLauncher.launch(intent);
+        });
+        AlertDialog dialog = builder.create();
+        dialog.show();
+        dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(this, com.sendbird.uikit.R.color.secondary_300));
     }
 }