+++ uikit-custom-sample/src/main/java/com/sendbird/uikit/customsample/openchannel/community/CreateCommunityActivity.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.text.Editable;
@@ -25,6 +24,7 @@ import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.sendbird.android.OpenChannel;
 import com.sendbird.android.OpenChannelParams;
 import com.sendbird.android.SendBird;
+import com.sendbird.uikit.SendBirdUIKit;
 import com.sendbird.uikit.customsample.R;
 import com.sendbird.uikit.customsample.consts.StringSet;
 import com.sendbird.uikit.customsample.databinding.ActivityCreateCommunityBinding;
@@ -45,7 +45,7 @@ import java.util.Locale;
 
 
 public class CreateCommunityActivity extends AppCompatActivity {
-    private String[] REQUIRED_PERMISSIONS;
+    private final String[] REQUIRED_PERMISSIONS = PermissionUtils.CAMERA_PERMISSION;
 
     private static final int STORAGE_PERMISSIONS_REQUEST_CODE = 1001;
     private static final int PERMISSION_SETTINGS_REQUEST_ID = 2000;
@@ -75,7 +75,7 @@ public class CreateCommunityActivity extends AppCompatActivity {
                 return;
             }
 
-            ActivityCompat.requestPermissions(this, REQUIRED_PERMISSIONS, STORAGE_PERMISSIONS_REQUEST_CODE);
+            requestPermission(REQUIRED_PERMISSIONS);
         });
         binding.etTitle.addTextChangedListener(new TextWatcher() {
             @Override
@@ -91,15 +91,7 @@ public class CreateCommunityActivity extends AppCompatActivity {
             public void afterTextChanged(Editable s) {}
         });
         binding.clearButton.setOnClickListener(v -> binding.etTitle.setText(""));
-
-        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-            REQUIRED_PERMISSIONS = new String[]{Manifest.permission.CAMERA,
-                    Manifest.permission.READ_EXTERNAL_STORAGE};
-        } else {
-            REQUIRED_PERMISSIONS = new String[]{Manifest.permission.CAMERA,
-                    Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                    Manifest.permission.READ_EXTERNAL_STORAGE};
-        }
+        SendBirdUIKit.connect(null);
     }
 
     private void createCommunityChannel() {
@@ -171,6 +163,7 @@ public class CreateCommunityActivity extends AppCompatActivity {
 
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if (requestCode == STORAGE_PERMISSIONS_REQUEST_CODE && grantResults.length == REQUIRED_PERMISSIONS.length) {
             boolean isAllGranted = true;
             for (int result : grantResults) {
@@ -182,26 +175,6 @@ public class CreateCommunityActivity extends AppCompatActivity {
 
             if (isAllGranted) {
                 showMediaSelectDialog();
-            } else {
-                String[] notGranted = PermissionUtils.getNotGrantedPermissions(this, permissions);
-                List<String> deniedList = PermissionUtils.getShowRequestPermissionRationale(this, permissions);
-                if (deniedList.size() == 0) {
-                    AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                    builder.setTitle(getString(com.sendbird.uikit.R.string.sb_text_dialog_permission_title));
-                    builder.setMessage(getPermissionGuildeMessage(this, notGranted[0]));
-                    builder.setPositiveButton(com.sendbird.uikit.R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
-                        Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                        intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                        intent.addCategory(Intent.CATEGORY_DEFAULT);
-                        intent.setData(Uri.parse("package:" + getPackageName()));
-                        intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
-                        intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
-                        startActivityForResult(intent, PERMISSION_SETTINGS_REQUEST_ID);
-                    });
-                    AlertDialog dialog = builder.create();
-                    dialog.show();
-                    dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(this, com.sendbird.uikit.R.color.secondary_300));
-                }
             }
         }
     }
@@ -224,7 +197,7 @@ public class CreateCommunityActivity extends AppCompatActivity {
                 if (key == com.sendbird.uikit.R.string.sb_text_channel_settings_change_channel_image_camera) {
                     takeCamera();
                 } else if (key == com.sendbird.uikit.R.string.sb_text_channel_settings_change_channel_image_gallery) {
-                    pickImage();
+                    takePhoto();
                 } else {
                     removeFile();
                 }
@@ -242,8 +215,8 @@ public class CreateCommunityActivity extends AppCompatActivity {
         }
     }
 
-    private void pickImage() {
-        Intent intent = IntentUtils.getGalleryIntent();
+    private void takePhoto() {
+        Intent intent = IntentUtils.getImageGalleryIntent();
         startActivityForResult(intent, PICK_IMAGE_ACTIVITY_REQUEST_CODE);
     }
 
@@ -253,7 +226,7 @@ public class CreateCommunityActivity extends AppCompatActivity {
         binding.ivChannelCover.setImageResource(0);
     }
 
-    private static String getPermissionGuildeMessage(@NonNull Context context, @NonNull String permission) {
+    private static String getPermissionGuideMessage(@NonNull Context context, @NonNull String permission) {
         int textResId;
         if (Manifest.permission.CAMERA.equals(permission)) {
             textResId = com.sendbird.uikit.R.string.sb_text_need_to_allow_permission_camera;
@@ -261,5 +234,41 @@ public class CreateCommunityActivity extends AppCompatActivity {
             textResId = com.sendbird.uikit.R.string.sb_text_need_to_allow_permission_storage;
         }
         return String.format(Locale.US, context.getString(textResId), ContextUtils.getApplicationName(context));
+    }
+
+    private void requestPermission(@NonNull String[] permissions) {
+        // 1. check permission
+        final boolean hasPermission = PermissionUtils.hasPermissions(this, permissions);
+        if (hasPermission) {
+            showMediaSelectDialog();
+            return;
+        }
+
+        // 2. determine whether rationale popup should show
+        final List<String> deniedList = PermissionUtils.getExplicitDeniedPermissionList(this, permissions);
+        if (!deniedList.isEmpty()) {
+            showPermissionRationalePopup(deniedList.get(0));
+            return;
+        }
+        // 3. request permission
+        ActivityCompat.requestPermissions(this, REQUIRED_PERMISSIONS, STORAGE_PERMISSIONS_REQUEST_CODE);
+    }
+
+    private void showPermissionRationalePopup(@NonNull String permission) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setTitle(getString(com.sendbird.uikit.R.string.sb_text_dialog_permission_title));
+        builder.setMessage(getPermissionGuideMessage(this, permission));
+        builder.setPositiveButton(com.sendbird.uikit.R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
+            Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.addCategory(Intent.CATEGORY_DEFAULT);
+            intent.setData(Uri.parse("package:" + getPackageName()));
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
+            intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
+            startActivityForResult(intent, PERMISSION_SETTINGS_REQUEST_ID);
+        });
+        AlertDialog dialog = builder.create();
+        dialog.show();
+        dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(this, com.sendbird.uikit.R.color.secondary_300));
     }
 }