+++ uikit-sample/src/main/java/com/sendbird/uikit_messaging_android/SettingsFragment.java
@@ -1,11 +1,12 @@
 package com.sendbird.uikit_messaging_android;
 
+import static android.app.Activity.RESULT_OK;
+
 import android.Manifest;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.view.LayoutInflater;
@@ -48,10 +49,8 @@ import java.util.List;
 import java.util.Locale;
 import java.util.TimeZone;
 
-import static android.app.Activity.RESULT_OK;
-
 public class SettingsFragment extends Fragment {
-    private String[] REQUIRED_PERMISSIONS;
+    private final String[] REQUIRED_PERMISSIONS = PermissionUtils.CAMERA_PERMISSION;
 
     private static final int STORAGE_PERMISSIONS_REQUEST_CODE = 1001;
     private static final int PERMISSION_SETTINGS_REQUEST_ID = 2000;
@@ -60,6 +59,12 @@ public class SettingsFragment extends Fragment {
 
     private FragmentSettingsBinding binding;
     private Uri mediaUri;
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        SendBirdUIKit.connect(null);
+    }
 
     @Nullable
     @Override
@@ -156,26 +161,6 @@ public class SettingsFragment extends Fragment {
 
             if (isAllGranted) {
                 showMediaSelectDialog();
-            } else {
-                String[] notGranted = PermissionUtils.getNotGrantedPermissions(getContext(), permissions);
-                List<String> deniedList = PermissionUtils.getShowRequestPermissionRationale(getActivity(), permissions);
-                if (deniedList.size() == 0 && getActivity() != null) {
-                    AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-                    builder.setTitle(getActivity().getString(com.sendbird.uikit.R.string.sb_text_dialog_permission_title));
-                    builder.setMessage(getPermissionGuildeMessage(getActivity(), notGranted[0]));
-                    builder.setPositiveButton(com.sendbird.uikit.R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
-                        Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                        intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                        intent.addCategory(Intent.CATEGORY_DEFAULT);
-                        intent.setData(Uri.parse("package:" + getActivity().getPackageName()));
-                        intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
-                        intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
-                        startActivityForResult(intent, PERMISSION_SETTINGS_REQUEST_ID);
-                    });
-                    AlertDialog dialog = builder.create();
-                    dialog.show();
-                    dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(getContext(), com.sendbird.uikit.R.color.secondary_300));
-                }
             }
         }
     }
@@ -183,15 +168,6 @@ public class SettingsFragment extends Fragment {
     private void initPage() {
         if (getContext() == null) {
             return;
-        }
-
-        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-            REQUIRED_PERMISSIONS = new String[]{Manifest.permission.CAMERA,
-                    Manifest.permission.READ_EXTERNAL_STORAGE};
-        } else {
-            REQUIRED_PERMISSIONS = new String[]{Manifest.permission.CAMERA,
-                    Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                    Manifest.permission.READ_EXTERNAL_STORAGE};
         }
 
         boolean useHeader = true;
@@ -313,7 +289,8 @@ public class SettingsFragment extends Fragment {
                     return;
                 }
 
-                requestPermissions(REQUIRED_PERMISSIONS, STORAGE_PERMISSIONS_REQUEST_CODE);
+                if (getActivity() == null) return;
+                requestPermission(REQUIRED_PERMISSIONS);
             }
         }).showSingle(getFragmentManager());
     }
@@ -407,7 +384,7 @@ public class SettingsFragment extends Fragment {
                         if (key == com.sendbird.uikit.R.string.sb_text_channel_settings_change_channel_image_camera) {
                             takeCamera();
                         } else if (key == com.sendbird.uikit.R.string.sb_text_channel_settings_change_channel_image_gallery) {
-                            pickImage();
+                            takePhoto();
                         }
                     } catch (Exception e) {
                         Logger.e(e);
@@ -427,12 +404,12 @@ public class SettingsFragment extends Fragment {
         }
     }
 
-    private void pickImage() {
-        Intent intent = IntentUtils.getGalleryIntent();
+    private void takePhoto() {
+        Intent intent = IntentUtils.getImageGalleryIntent();
         startActivityForResult(intent, PICK_IMAGE_ACTIVITY_REQUEST_CODE);
     }
 
-    private static String getPermissionGuildeMessage(@NonNull Context context, @NonNull String permission) {
+    private static String getPermissionGuideMessage(@NonNull Context context, @NonNull String permission) {
         int textResId;
         if (Manifest.permission.CAMERA.equals(permission)) {
             textResId = com.sendbird.uikit.R.string.sb_text_need_to_allow_permission_camera;
@@ -445,5 +422,42 @@ public class SettingsFragment extends Fragment {
     protected boolean isActive() {
         boolean isDeactivated = isRemoving() || isDetached() || getContext() == null;
         return !isDeactivated;
+    }
+
+    private void requestPermission(@NonNull String[] permissions) {
+        if (getContext() == null || getActivity() == null) return;
+        // 1. check permission
+        final boolean hasPermission = PermissionUtils.hasPermissions(getContext(), permissions);
+        if (hasPermission) {
+            showMediaSelectDialog();
+            return;
+        }
+
+        // 2. determine whether rationale popup should show
+        final List<String> deniedList = PermissionUtils.getExplicitDeniedPermissionList(getActivity(), permissions);
+        if (!deniedList.isEmpty()) {
+            showPermissionRationalePopup(deniedList.get(0));
+            return;
+        }
+        // 3. request permission
+        requestPermissions(REQUIRED_PERMISSIONS, STORAGE_PERMISSIONS_REQUEST_CODE);
+    }
+
+    private void showPermissionRationalePopup(@NonNull String permission) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(requireContext());
+        builder.setTitle(getString(com.sendbird.uikit.R.string.sb_text_dialog_permission_title));
+        builder.setMessage(getPermissionGuideMessage(requireContext(), permission));
+        builder.setPositiveButton(com.sendbird.uikit.R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
+            Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.addCategory(Intent.CATEGORY_DEFAULT);
+            intent.setData(Uri.parse("package:" + requireContext().getPackageName()));
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
+            intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
+            startActivityForResult(intent, PERMISSION_SETTINGS_REQUEST_ID);
+        });
+        AlertDialog dialog = builder.create();
+        dialog.show();
+        dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(requireContext(), com.sendbird.uikit.R.color.secondary_300));
     }
 }