+++ uikit/src/main/java/com/sendbird/uikit/fragments/ChannelSettingsFragment.java
@@ -44,6 +44,7 @@ import com.sendbird.uikit.tasks.TaskQueue;
 import com.sendbird.uikit.utils.DialogUtils;
 import com.sendbird.uikit.utils.FileUtils;
 import com.sendbird.uikit.utils.IntentUtils;
+import com.sendbird.uikit.utils.PermissionUtils;
 import com.sendbird.uikit.utils.TextUtils;
 import com.sendbird.uikit.widgets.ChannelSettingsView;
 
@@ -207,18 +208,13 @@ public class ChannelSettingsFragment extends BaseFragment implements PermissionF
     }
 
     private boolean isCurrentChannel(@NonNull String channelUrl) {
+        if (channel == null) return false;
         return channelUrl.equals(channel.getUrl());
     }
 
     @Override
     public String[] getPermissions(int requestCode) {
-        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-            return new String[]{Manifest.permission.CAMERA,
-                    Manifest.permission.READ_EXTERNAL_STORAGE};
-        }
-        return new String[]{Manifest.permission.CAMERA,
-                Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                Manifest.permission.READ_EXTERNAL_STORAGE};
+        return PermissionUtils.CAMERA_PERMISSION;
     }
 
     @Override
@@ -397,7 +393,7 @@ public class ChannelSettingsFragment extends BaseFragment implements PermissionF
     }
 
     private void pickImage() {
-        Intent intent = IntentUtils.getGalleryIntent();
+        Intent intent = IntentUtils.getImageGalleryIntent();
         startActivityForResult(intent, PICK_IMAGE_PERMISSIONS_REQUEST_CODE);
     }