+++ uikit/src/main/java/com/sendbird/uikit/fragments/OpenChannelFragment.java
@@ -69,6 +69,7 @@ import com.sendbird.uikit.utils.DialogUtils;
 import com.sendbird.uikit.utils.FileUtils;
 import com.sendbird.uikit.utils.IntentUtils;
 import com.sendbird.uikit.utils.MessageUtils;
+import com.sendbird.uikit.utils.PermissionUtils;
 import com.sendbird.uikit.utils.SoftInputUtils;
 import com.sendbird.uikit.utils.TextUtils;
 import com.sendbird.uikit.vm.FileDownloader;
@@ -583,13 +584,7 @@ public class OpenChannelFragment extends BaseOpenChannelFragment implements OnId
         checkPermission(PERMISSION_REQUEST_ALL, new IPermissionHandler() {
             @Override
             public String[] getPermissions(int requestCode) {
-                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-                    return new String[]{Manifest.permission.CAMERA,
-                            Manifest.permission.READ_EXTERNAL_STORAGE};
-                }
-                return new String[]{Manifest.permission.CAMERA,
-                        Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                        Manifest.permission.READ_EXTERNAL_STORAGE};
+                return PermissionUtils.CAMERA_PERMISSION;
             }
 
             @Override
@@ -613,11 +608,7 @@ public class OpenChannelFragment extends BaseOpenChannelFragment implements OnId
         checkPermission(PERMISSION_REQUEST_STORAGE, new IPermissionHandler() {
             @Override
             public String[] getPermissions(int requestCode) {
-                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-                    return new String[]{Manifest.permission.READ_EXTERNAL_STORAGE};
-                }
-                return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                        Manifest.permission.READ_EXTERNAL_STORAGE};
+                return PermissionUtils.GET_CONTENT_PERMISSION;
             }
 
             @Override
@@ -638,11 +629,7 @@ public class OpenChannelFragment extends BaseOpenChannelFragment implements OnId
         checkPermission(PERMISSION_REQUEST_STORAGE, new IPermissionHandler() {
             @Override
             public String[] getPermissions(int requestCode) {
-                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
-                    return new String[]{Manifest.permission.READ_EXTERNAL_STORAGE};
-                }
-                return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                        Manifest.permission.READ_EXTERNAL_STORAGE};
+                return PermissionUtils.GET_CONTENT_PERMISSION;
             }
 
             @Override
@@ -1051,8 +1038,7 @@ public class OpenChannelFragment extends BaseOpenChannelFragment implements OnId
                 @Override
                 @NonNull
                 public String[] getPermissions(int requestCode) {
-                    return new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                            Manifest.permission.READ_EXTERNAL_STORAGE};
+                    return PermissionUtils.GET_CONTENT_PERMISSION;
                 }
 
                 @Override