+++ uikit/src/main/java/com/sendbird/uikit/fragments/PermissionFragment.java
@@ -1,5 +1,7 @@
 package com.sendbird.uikit.fragments;
 
+import static android.app.Activity.RESULT_OK;
+
 import android.Manifest;
 import android.app.Activity;
 import android.content.Context;
@@ -7,16 +9,16 @@ import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.content.pm.PermissionGroupInfo;
 import android.content.pm.PermissionInfo;
-import android.content.res.Resources;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
-import android.os.Build;
 import android.provider.Settings;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
+import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
+import androidx.core.content.res.ResourcesCompat;
 import androidx.fragment.app.Fragment;
 
 import com.sendbird.uikit.R;
@@ -25,8 +27,6 @@ import com.sendbird.uikit.utils.PermissionUtils;
 
 import java.util.List;
 import java.util.Locale;
-
-import static android.app.Activity.RESULT_OK;
 
 public abstract class PermissionFragment extends Fragment {
     interface IPermissionHandler {
@@ -39,16 +39,31 @@ public abstract class PermissionFragment extends Fragment {
     private final int PERMISSION_SETTINGS_REQUEST_ID = 100;
 
     void checkPermission(int requestCode, @NonNull IPermissionHandler handler) {
+        if (getContext() == null || getActivity() == null) return;
         this.requestCode = requestCode;
         this.handler = handler;
         final String[] permissions = handler.getPermissions(requestCode);
+        // 1. check permission
         final boolean hasPermission = PermissionUtils.hasPermissions(getContext(), permissions);
         if (hasPermission) {
-            handler.onPermissionGranted(requestCode);
+            notifyPermissionResult(handler);
+            return;
+        }
+
+        // 2. determine whether rationale popup should show
+        final List<String> deniedList = PermissionUtils.getExplicitDeniedPermissionList(getActivity(), permissions);
+        if (!deniedList.isEmpty()) {
+            showPermissionRationalePopup(deniedList.get(0));
             return;
         }
 
         requestPermissions(permissions, requestCode);
+    }
+
+    private void notifyPermissionResult(@Nullable IPermissionHandler handler) {
+        if (handler != null) {
+            handler.onPermissionGranted(requestCode);
+        }
     }
 
     @Override
@@ -64,31 +79,7 @@ public abstract class PermissionFragment extends Fragment {
             }
 
             if (isAllGranted) {
-                handler.onPermissionGranted(requestCode);
-            } else {
-                String[] notGranted = PermissionUtils.getNotGrantedPermissions(getContext(), permissions);
-                List<String> deniedList = PermissionUtils.getShowRequestPermissionRationale(getActivity(), permissions);
-                if (deniedList != null && deniedList.size() == 0) {
-                    Drawable icon = getPermissionDrawable(getActivity(), notGranted[0]);
-                    AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-                    builder.setTitle(getActivity().getString(R.string.sb_text_dialog_permission_title));
-                    builder.setMessage(getPermissionGuideMessage(getActivity(), notGranted[0]));
-                    if (icon != null) {
-                        builder.setIcon(icon);
-                    }
-                    builder.setPositiveButton(R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
-                        Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                        intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                        intent.addCategory(Intent.CATEGORY_DEFAULT);
-                        intent.setData(Uri.parse("package:" + getActivity().getPackageName()));
-                        intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
-                        intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
-                        startActivityForResult(intent, PERMISSION_SETTINGS_REQUEST_ID);
-                    });
-                    AlertDialog dialog = builder.create();
-                    dialog.show();
-                    dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(getContext(), R.color.secondary_300));
-                }
+                notifyPermissionResult(handler);
             }
         }
     }
@@ -101,38 +92,56 @@ public abstract class PermissionFragment extends Fragment {
                 String[] permissions = handler.getPermissions(this.requestCode);
                 final boolean hasPermission = PermissionUtils.hasPermissions(getContext(), permissions);
                 if (hasPermission) {
-                    handler.onPermissionGranted(requestCode);
+                    handler.onPermissionGranted(this.requestCode);
                 }
             }
         }
     }
 
-    private static Drawable getPermissionDrawable(Activity activity, String permission) {
+    private void showPermissionRationalePopup(@NonNull String permission) {
+        Drawable icon = getPermissionDrawable(requireActivity(), permission);
+        AlertDialog.Builder builder = new AlertDialog.Builder(requireContext());
+        builder.setTitle(requireContext().getString(R.string.sb_text_dialog_permission_title));
+        builder.setMessage(getPermissionGuideMessage(requireContext(), permission));
+        if (icon != null) {
+            builder.setIcon(icon);
+        }
+        builder.setPositiveButton(R.string.sb_text_go_to_settings, (dialogInterface, i) -> {
+            Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+            intent.addCategory(Intent.CATEGORY_DEFAULT);
+            intent.setData(Uri.parse("package:" + requireContext().getPackageName()));
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
+            intent.addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS);
+            startActivityForResult(intent, PERMISSION_SETTINGS_REQUEST_ID);
+        });
+        AlertDialog dialog = builder.create();
+        dialog.show();
+        dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(ContextCompat.getColor(requireContext(), R.color.secondary_300));
+    }
+
+    @Nullable
+    private static Drawable getPermissionDrawable(@NonNull Activity activity, @NonNull String permission) {
         Drawable drawable = null;
         try {
             PackageManager pm = activity.getPackageManager();
             PermissionInfo permissionInfo = pm.getPermissionInfo(permission, 0);
+            if (permissionInfo.group == null) return null;
             PermissionGroupInfo groupInfo = pm.getPermissionGroupInfo(permissionInfo.group, 0);
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                drawable = pm.getResourcesForApplication("android").getDrawable(groupInfo.icon, activity.getTheme());
-            } else {
-                drawable = pm.getResourcesForApplication("android").getDrawable(groupInfo.icon);
-            }
-        } catch (PackageManager.NameNotFoundException | Resources.NotFoundException e) {
+            drawable = ResourcesCompat.getDrawable(pm.getResourcesForApplication("android"), groupInfo.icon, activity.getTheme());
+        } catch (Exception e) {
+            //Logger.w(e);
         }
         return drawable;
     }
 
+    @NonNull
     private static String getPermissionGuideMessage(@NonNull Context context, @NonNull String permission) {
         int textResId;
-        switch (permission) {
-            case Manifest.permission.CAMERA:
-                textResId = R.string.sb_text_need_to_allow_permission_camera;
-                break;
-            default:
-                textResId = R.string.sb_text_need_to_allow_permission_storage;
-                break;
-
+        if (Manifest.permission.CAMERA.equals(permission)) {
+            textResId = R.string.sb_text_need_to_allow_permission_camera;
+        } else {
+            textResId = R.string.sb_text_need_to_allow_permission_storage;
         }
         return String.format(Locale.US, context.getString(textResId), ContextUtils.getApplicationName(context));
     }