+++ uikit/src/main/java/com/sendbird/uikit/utils/UserUtils.java
@@ -33,10 +33,17 @@ public class UserUtils {
 
     @NonNull
     public static String getDisplayName(@NonNull Context context, @Nullable User user) {
+        return getDisplayName(context, user, false);
+    }
+
+    @NonNull
+    public static String getDisplayName(@NonNull Context context, @Nullable User user, boolean usePronouns) {
         String nickname = context.getString(R.string.sb_text_channel_list_title_unknown);
         if (user == null) return nickname;
 
-        if (user.getUserId() != null && user.getUserId().equals(SendBird.getCurrentUser().getUserId())) {
+        if (usePronouns && user.getUserId() != null &&
+                SendBird.getCurrentUser() != null &&
+                user.getUserId().equals(SendBird.getCurrentUser().getUserId())) {
             nickname = context.getString(R.string.sb_text_you);
         } else if (!TextUtils.isEmpty(user.getNickname())) {
             nickname = user.getNickname();